+++ .gitignore
@@ -10,6 +10,7 @@ coverage/
 *.stackdump
 .nyc_output/
 coverage.lcov
+documentation/
 
 # Ignore the english compiled file
 i18n/resources.en.json
