+++ src/client/core/properties/index.ts
@@ -1,0 +1,1 @@
+export * from "./batch-explorer-properties";
