+++ src/client/core/unique-window.ts
@@ -1,13 +1,16 @@
 import { log } from "@batch-flask/utils";
+import { BatchExplorerProperties } from "client/core/properties";
 import { BatchExplorerApplication } from ".";
 
 export abstract class GenericWindow {
     public expectedClose = false;
     public domReady: Promise<void>;
+    protected properties: BatchExplorerProperties;
     protected _window: Electron.BrowserWindow;
     private _resolveDomReady: () => void;
 
     constructor(protected batchExplorerApp: BatchExplorerApplication) {
+        this.properties = this.batchExplorerApp.properties;
         this.domReady = new Promise((resolve) => {
             this._resolveDomReady = resolve;
         });
