+++ src/client/menu/index.ts
@@ -1,0 +1,1 @@
+export * from "./main-application-menu";
