+++ src/client/menu/menu.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from "@angular/core";
+import { MainApplicationMenu } from "./main-application-menu";
+
+@NgModule({
+    providers: [
+        MainApplicationMenu,
+    ],
+})
+export class MenuModule {
+}
